requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0

# AI & Machine Learning Dependencies - Part 2
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-groq==0.0.1
langchain-community==0.0.13
langchain-core==0.1.10
chromadb==0.4.22
sentence-transformers==2.2.2
